Lag filstørrelse-konstant
